Enabled conversion of geometry data to double precision for XDMF files.
XDM assumes geometry is double precision. XDMF does not have this requirement.
Therefore, the XDMF plugin converts geometry values to double precision.

+++ src/plugins/xdmf/test/TestXmfReader.cpp
@@ -73,16 +73,18 @@ xdm::RefPtr< xdmGrid::UniformGrid > build2DGrid() {
   }
 
   // Geometry
-  xdm::RefPtr< xdm::VectorStructuredArray< double > > meshValues[2];
+  xdm::RefPtr< xdm::VectorStructuredArray< float > > meshValues[2];
   {
     xdm::RefPtr< xdmGrid::TensorProductGeometry > geometry(
       new xdmGrid::TensorProductGeometry( 2 ) );
     for ( int i = 0; i < 2; i++ ) {
+      // Build the Geometry data as single precision to force an up conversion
+      // upon read.
       xdm::RefPtr< xdm::UniformDataItem > dataItem( new xdm::UniformDataItem );
-      dataItem->setDataType( xdm::primitiveType::kDouble );
+      dataItem->setDataType( xdm::primitiveType::kFloat );
       dataItem->setDataspace( xdm::makeShape( kMeshSize[i] ) );
       meshValues[i] = xdm::makeRefPtr(
-        new xdm::VectorStructuredArray< double >( kMeshSize[i] ) );
+        new xdm::VectorStructuredArray< float >( kMeshSize[i] ) );
       for ( int j = 0; j < kMeshSize[i]; j++ ) {
         (*meshValues[i])[j] = kRange[i][0] + j * ( (kRange[i][1] - kRange[i][0]) / kMeshSize[i] );
       }
